Take into account also reads from Cache
